Fix babel, widget directory structure
