+++ package.json
@@ -17,6 +17,10 @@
     "url": "https://github.com/KC3Kai/kc3-firefox/issues"
   },
   "devDependencies": {
+    "@babel/core": "^7.0.0-beta.32",
+    "@babel/preset-env": "^7.0.0-beta.32",
+    "babel-loader": "^8.0.0-beta.0",
+    "babel-plugin-syntax-dynamic-import": "^6.18.0",
     "clean-webpack-plugin": "^0.1.17",
     "copy-webpack-plugin": "^4.2.1",
     "css-loader": "^0.28.7",
